misc
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@10159 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ WebPortalResponse.mm
@@ -340,7 +340,7 @@
 		NSString* evaldToken = [self evaluateToken:token withDictionary:dict context:context mustReevaluate:mustReevaluate];
 		
 		if ([part1 isEqualToString:@"ZWS"])
-			evaldToken = [[evaldToken componentsWithLength:1] componentsJoinedByString:[NSString stringWithFormat:@"%C",0x200b]];
+			evaldToken = [[evaldToken componentsWithLength:1] componentsJoinedByString:[NSString stringWithFormat:@"%C", (unsigned short)0x200b]];
 		
 		return [evaldToken xmlEscapedString];
 	}
